% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyFunctions.r
\name{add_adduct}
\alias{add_adduct}
\title{add adduct m/z to metabolite}
\usage{
add_adduct(lib, type = "RP_pos")
}
\arguments{
\item{lib}{spectrum library, A list}

\item{type}{column and ionization mode, could be one of c("RP_pos","RP_neg","HILIC_pos", "HILIC_neg")}
}
\value{
a matrix, represents each metablites adduct m/z
}
\description{
add adduct m/z to metabolite
}
\examples{
\donttest{
  data("spectrumDB", package="RFQI")
  data("adductDB", package="RFQI")
  add_adduct(lib=spectrumDB, type="RP_pos")
}
}
