% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adductDB}
\alias{adductDB}
\title{possible adduct when ionization}
\format{
A list containing four matrix, each matrix is an adduct table
\describe{
  \item{HILIC_pos}{chromatographic column is HILIC, and ionization mode is "positive"}
  \item{HILIC_neg}{chromatographic column is HILIC, and ionization mode is "negative"}
  \item{RP_pos}{chromatographic column is RP, and ionization mode is "positive"}
  \item{RP_neg}{chromatographic column is RP, and ionization mode is "negative"}
}
}
\source{
data("adductDB", package="RFQI")
}
\usage{
adductDB
}
\description{
possible adduct when ionization
}
\keyword{datasets}
