% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMS2functions.r
\name{align_MS2_to_MS1}
\alias{align_MS2_to_MS1}
\title{alignMS2_to_MS1}
\usage{
align_MS2_to_MS1(ms2Info, features, cores = 3)
}
\arguments{
\item{ms2Info}{result of getMS2FromFiles}

\item{features}{result of getPeakRange_multipleFiles}

\item{cores}{numeric, for parallel computing}
}
\value{
list
}
\description{
alignMS2_to_MS2 is to align MS2 database to MS1 index according to ["mzmin","mzmax","rtmin","rtmax"]
}
