% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateFTFunctions.r
\name{checkFT}
\alias{checkFT}
\title{Extract peaks mapping to certain feature}
\usage{
checkFT(features, idx)
}
\arguments{
\item{features}{list results from function getPeaks_multipleFiles, containing two data frame: peaks and features}

\item{idx}{index of features}
}
\value{
data frame, peaks corresponding idx feature
}
\description{
Extract peaks mapping to certain feature
}
