% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2ClusterFunctions.r
\name{checkMS2}
\alias{checkMS2}
\title{Extract MS2 mapping to certain Feature}
\usage{
checkMS2(idx, MS2DB)
}
\arguments{
\item{idx}{index of features}

\item{MS2DB}{db.MS2 object}
}
\value{
list, similar to db.MS2, but only contains MS2 and precursorInfo of corresponding idx
}
\description{
Extract MS2 mapping to certain Feature
}
