% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{chrom_plot}
\alias{chrom_plot}
\title{plot EIC}
\usage{
chrom_plot(f.in, refFile = NULL, features, idx, mzr = NULL,
  rtr = NULL, col = NULL)
}
\arguments{
\item{f.in}{path of mzxml files}

\item{refFile}{path of reference mzXML files, using for retention time alignment}

\item{features}{results from function getPeaks_multipleFiles, containing two data frame: raw peaks and features}

\item{idx}{id of features from featureTable, chrom_plot only plot idx corresponding EIC}

\item{mzr}{m/z range, if appointed, plot mzr area EIC}

\item{rtr}{rt range, if appointed, plot rtr area EIC}

\item{col}{colours for each sample, if not appointed, using brewer.pal(9, "Set3")}
}
\value{
a plot
}
\description{
plot EIC
}
