% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFTfunctions.r
\name{convert_char_to_num}
\alias{convert_char_to_num}
\title{ID conversion}
\usage{
convert_char_to_num(x, prefix = "FT")
}
\arguments{
\item{x}{character}

\item{prefix}{prefix of x}
}
\value{
a numeric
}
\description{
convert a character to a numeric
}
\examples{
convert_char_to_num("FT0001", prefix="FT")
}
