% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFTfunctions.r
\name{convert_num_to_char}
\alias{convert_num_to_char}
\title{ID conversion, used to encoding rownames and colnames}
\usage{
convert_num_to_char(x, prefix = "FT", n = 4)
}
\arguments{
\item{x}{numeric}

\item{prefix}{character, this will be added to the head of x}

\item{n}{the result length of character}
}
\value{
a character whose length is \code{n}, and with a \code{prefix}
}
\description{
convert a number into a character with a prefix in fixed length
}
\examples{
convert_num_to_char(1, prefix="FT", n=4)
}
