% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db.MS2}
\alias{db.MS2}
\title{MS2 spectrum Mapping to features}
\format{A list containing two elements
\describe{
  \item{MS2}{MS2 spectrum}
  \item{presursorInfo}{precursor information corresponding MS2 spectrum}
}}
\source{
data("db.MS2", package="RFQI")
}
\usage{
db.MS2
}
\description{
MS2 spectrum Mapping to features
}
\keyword{datasets}
