% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFTfunctions.r
\name{errorBar}
\alias{errorBar}
\title{add error bar for boxplot}
\usage{
errorBar(x0 = 1, y0, x1 = 2, y1, y, pvalue)
}
\arguments{
\item{x0}{numeric}

\item{y0}{numeric}

\item{x1}{numeric}

\item{y1}{numeric}

\item{y}{numeric}

\item{pvalue}{character}
}
\value{
null
}
\description{
add error bar for boxplot
}
