% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{features}
\alias{features}
\title{Features resulting from mzXML files in extdata dir}
\format{A list containing two data frame
\describe{
  \item{group_FT}{features after combining peaks}
  \item{peaks}{peaks extracted from mzXML files using CentWave method}
}}
\source{
data("features", package="RFQI")
}
\usage{
features
}
\description{
Features resulting from mzXML files in extdata dir
}
\keyword{datasets}
