% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFTfunctions.r
\name{findPeaks_L}
\alias{findPeaks_L}
\title{get peaks from multiple samples(f.in)}
\usage{
findPeaks_L(f.in, ref = NULL, param = new("ParamSet"))
}
\arguments{
\item{f.in}{a vertor contains paths of mzxml files}

\item{ref}{Path of reference sample for rentention time alignment}

\item{param}{ParamSet class, must contains ppm, peakwidth, and noise, these three parameter are for peak detection Using Centwave method}
}
\value{
if has reference sample(ref), return RT-correction peaks(["mzmin","mzmax","rtmin","rtmax"])
}
\description{
detect peaks from multiple sample by centwave method
}
\examples{
files <- dir(system.file(package="RFQI", dir="extdata"), 
             full.name=TRUE, 
             pattern="mzXML$")
param = new("ParamSet", binSize=0.25)
peaks = findPeaks_L(f.in=files, ref=NULL, param=param)
}
