% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateFTFunctions.r
\name{generate_feature_from_MS2}
\alias{generate_feature_from_MS2}
\title{generate_feature_from_MS2 is to assemble MS2's precursor into features}
\usage{
generate_feature_from_MS2(db.MS2, features, param = new("ParamSet"))
}
\arguments{
\item{db.MS2}{results from function align_MS2_to_MS1}

\item{features}{results from getPeaks_multipleFiles}

\item{param}{ParamSet object}
}
\value{
list as features
}
\description{
generate_feature_from_MS2 is to assemble MS2's precursor into features
}
