% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFTfunctions.r
\name{get_compareFT}
\alias{get_compareFT}
\title{get_compareFT}
\usage{
get_compareFT(M1, M2, absMz = 0.005, absRt = 15, cores = 2)
}
\arguments{
\item{M1}{first feature table}

\item{M2}{second feature table}

\item{absMz}{numeric}

\item{absRt}{numeric}

\item{cores}{thread for parallel computing}
}
\value{
list
}
\description{
compare two feature table overlap
}
