% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyFunctions.r
\name{get_identify}
\alias{get_identify}
\title{Identify features based on multiple MS2 similarity}
\usage{
get_identify(lib, MS2DB, MS2_inner_cor, MS1_idx = NULL, absMz = 0.015,
  adduct = "RP_pos", cores = 1)
}
\arguments{
\item{lib}{standard metabolite MS2 spectrum database, We have offer a database containing 3000 metabolites, using \code{data("spectrumDB", package="RFQI")}}

\item{MS2DB}{db.MS2 object, result from function align_MS2_to_MS1}

\item{MS2_inner_cor}{list containing ms2 similarity score matrix}

\item{MS1_idx}{if not NULL, only identify appoint MS1_index, else identify all features having MS2}

\item{absMz}{mz tolerance in Da unit, this is used for find matched precursor}

\item{adduct}{possible adduct when ionization, can be one of c("RP_pos", "RP_neg", "HILIC_pos", "HILIC_neg"), RP represents reverse column, pos is positive mode}

\item{cores}{cores for parallel computing}
}
\value{
identification result
}
\description{
Need to prepare standard metabolites MS2 spectrum database
}
