% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateFTFunctions.r
\name{get_MS2_cor}
\alias{get_MS2_cor}
\title{Compute similarity score of two sets of MS2 spectrum}
\usage{
get_MS2_cor(MS2_set1, MS2_set2, cores = 2)
}
\arguments{
\item{MS2_set1}{list containing multiple MS2 spectrum}

\item{MS2_set2}{list containing multiple MS2 spectrum}

\item{cores}{threads for parallel computing}
}
\value{
similarity matrix
}
\description{
Compute similarity score of two sets of MS2 spectrum
}
