% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ms2cluster.r
\name{get_ms2cluster}
\alias{get_ms2cluster}
\title{Do MS2 clustering using dbscan}
\usage{
get_ms2cluster(idx, MS2_inner_cor, eps = 0.5, minPts = 2)
}
\arguments{
\item{idx}{index of features}

\item{MS2_inner_cor}{list, names are feature index, and each element is a MS2 similarity matrix}

\item{eps}{similarity score threshold}

\item{minPts}{minimum member of a node}
}
\value{
a list, cluster contains each cluster's member, and noise are isolate MS2
}
\description{
Do MS2 clustering using dbscan
}
