% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2SimilarityFunctions.r
\name{get_multiple_ms2Cor}
\alias{get_multiple_ms2Cor}
\alias{get_MS2_cor_inner}
\title{compute pair cosine similarity score of multiple MS2}
\usage{
get_multiple_ms2Cor(MS2_set, cores = 2, ppm = 30, sn = 3)
}
\arguments{
\item{MS2_set}{a list whose elements are MS2 spectrum}

\item{cores}{numeric for parallel computing}

\item{ppm}{m/z tolerance}

\item{sn}{signal to noise ratio}
}
\value{
a similarity score matrix
}
\description{
compute pair cosine similarity score of multiple MS2
}
