% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFTfunctions.r
\name{get_splited_features}
\alias{get_splited_features}
\title{split over-grouped features generated by getPeakRange_multipleFiles}
\usage{
get_splited_features(
  idx,
  features,
  param = new("ParamSet"),
  ns = 0.1,
  intMethod = "maxo"
)
}
\arguments{
\item{idx}{feauture id in feature tabel}

\item{features}{list containing group_FT(feature table) and peaks(original peaks)}

\item{param}{ParamSet object same to that of getPeakRange_multipleFiles}

\item{ns}{noise to signal ratio, peaks whose "maxo" or "into" less than ns will be discarded}

\item{intMethod}{method to evalute peaks intensity, one of c("maxo", "into")}
}
\value{
re-grouped feature table corresponding to idx
}
\description{
get_splited_features is to split over-grouped features by generated by getPeakRange_multipleFiles through reducing low intensity peaks and re-grouping
}
