% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2SimilarityFunctions.r
\name{GetDiffMZppm}
\alias{GetDiffMZppm}
\title{translate Da to ppm}
\usage{
GetDiffMZppm(mz, mz.ppm.thr = NULL)
}
\arguments{
\item{mz}{numeric vector, m/z values sorted by inreasing}

\item{mz.ppm.thr}{numeric, threshold of ppm difference}
}
\value{
vector, ppm difference of m/z with the first m/z
}
\description{
Calculating difference of two m/z, then convert diff into ppm
}
\examples{
\donttest{
  GetDiffMZppm(c(100.01,100.02, 100.03))
}
}
