% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyFunctions.r
\name{getFDR}
\alias{getFDR}
\title{Compute FDR}
\usage{
getFDR(score, ms2_inner_score, ratio = 0.1)
}
\arguments{
\item{score}{similarity vector of one spectrum with multiple MS2}

\item{ms2_inner_score}{similarity matrix of multiple MS2}

\item{ratio}{ratio threshold, is the minimum possibility of feature generating certain MS2}
}
\value{
ratio
}
\description{
Compute possibility of one features generates certain (often) MS2, if possibility greater than a threshold, we say this feature is the metabolite
}
