% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2ClusterFunctions.r
\name{getFeatureHasMS2}
\alias{getFeatureHasMS2}
\title{Get feature index that has MS2}
\usage{
getFeatureHasMS2(MS2DB, n = 2)
}
\arguments{
\item{MS2DB}{db.MS2 object}

\item{n}{minumum number of feature should have}
}
\value{
feature index that has MS2
}
\description{
Get feature index that has MS2
}
\examples{
\donttest{
  data("db.MS2", package="RFQI")
  getMS2Feature(db.MS2)
}
}
