% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFTfunctions.r
\name{getFileName}
\alias{getFileName}
\title{get file name from fll path}
\usage{
getFileName(path)
}
\arguments{
\item{path}{full path to a file containing format suffix}

\item{prefix}{character, this will be added to the head of x}

\item{n}{the result length of character}
}
\value{
filename
}
\description{
get file name from full path, e.g. "/A/B/C/D.txt" generates "D"
}
