% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2SimilarityFunctions.r
\name{GetMatchResult}
\alias{GetMatchResult}
\title{compute cosine similarity of two MS2 spectrum}
\usage{
GetMatchResult(spec.exp, spec.lib, sn = 3, ppm = 30)
}
\arguments{
\item{spec.exp}{experimental MS2 spectrum, columns are "mz" and "intensity"}

\item{spec.lib}{reference MS2 spectrum, columns are "mz" and "intensity"}

\item{sn}{signal to noise ratio}

\item{ppm}{m/z tolerence of MS2}
}
\value{
cosine similarity of spec.exp and spec.lib
}
\description{
compute cosine similarity of two MS2 spectrum
}
