% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMS2functions.r
\name{getMS2FromFiles}
\alias{getMS2FromFiles}
\title{getMS2FromFiles}
\usage{
getMS2FromFiles(files, ref_file = NULL)
}
\arguments{
\item{files}{path of mzxml files}

\item{ref_file}{path of referece mzxml file, using to do retention time alignment}
}
\value{
list, MS2 is a list containing MS2 spectrum; precursorInfo is a dataframe, the ith row of precursorInfo is the ith MS2's precursor information
}
\description{
get MS2 from multiple mzXML files
}
