% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFTfunctions.r
\name{getPeakRange_multipleFiles}
\alias{getPeakRange_multipleFiles}
\title{group peaks from multiple files}
\usage{
getPeakRange_multipleFiles(f.in = NULL, peaks = NULL, ref = NULL,
  param = new("ParamSet"), cores = 2, prefix = "FT")
}
\arguments{
\item{f.in}{input}

\item{peaks}{data frame}

\item{ref}{character}

\item{param}{ParamSet object}

\item{cores}{mc.cores}

\item{prefix}{prefix}
}
\value{
feature table
}
\description{
getPeakRange_multipleFiles is to group peaks from multiple files or a given peaks table
}
