% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMS2functions.r
\name{getPeaks_L}
\alias{getPeaks_L}
\title{Extract spectrum from mzXML files}
\usage{
getPeaks_L(
  file,
  peak_range = NULL,
  ref_file = NULL,
  MS1 = TRUE,
  step = 0.1,
  MS2 = FALSE
)
}
\arguments{
\item{file}{path of one mzxml file}

\item{peak_range}{dataframe, each row reprensents a peak/feature area, commonly result from \code{getPeakRange_multiples}}

\item{ref_file}{path of referece mzxml file, using to do retention time alignment}

\item{MS1}{logical, if TRUE, extract TIC intensity of peaks in peak_range}

\item{step}{numeric, use for integration}

\item{MS2}{logical, if TRUE, extract MS2 spectrum}
}
\value{
list
}
\description{
getPeaks_L is to get peaks and/or MS2 based on peak_range
if have ref_file, peaks and MS2 is RT-corrected
}
\examples{
\donttest{
files = files <- dir(system.file(package="RFQI", dir="extdata"), full.name=TRUE, pattern="mzXML$")
data("features", package="RFQI")
peak_range = features$group_FT
getPeaks_L(file=files[1], ref_file=files[2], peak_range=peak_range, MS1=TRUE, step=0.1, MS2=FALSE)
} 
 
}
