% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2SimilarityFunctions.r
\name{GetSpec2Match}
\alias{GetSpec2Match}
\title{combine two MS2 spectrum based on m/z}
\usage{
GetSpec2Match(spec.exp, spec.lib, ppm.ms2match = 30)
}
\arguments{
\item{spec.exp}{MS2 spectrum}

\item{spec.lib}{MS2 spectrum}

\item{ppm.ms2match}{m/z tolerance of MS2}
}
\value{
list, contains m/z alignment exp and lib spectrum
}
\description{
combine two MS2 spectrum based on m/z
}
