% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2ClusterFunctions.r
\name{getTypeMS2}
\alias{getTypeMS2}
\title{Select typical MS2 based on similarity matrix}
\usage{
getTypeMS2(MS2DB, corMatrix, idx = NULL, include.all = TRUE)
}
\arguments{
\item{MS2DB}{db.MS2 object}

\item{corMatrix}{list resulting from get_ms2Cor_inner, each element is a MS2 similarity Matrix}

\item{idx}{if appointed, only extract typical MS2 for features in idx}

\item{include.all}{logical, whether extract typical MS2 for features not in corMatrix (usually contains only 1 or 2 MS2)}
}
\value{
list, each element contains precursors' m/z and typical MS2 spectrum
}
\description{
Difference selecting typical MS2 spectrum in highest intensity or consensus MS2, here we define the typical MS2
is the center MS2 in similarity matrix
}
\examples{
\donttest{
data("db.MS2", package="RFQI")
corMatrix = get_ms2Cor_inner(MS2DB=db.MS2, n=3, ppm=30, sn=0, cores1=2)
MS2_with_mz = getTypeMS2(MS2DB = db.MS2, corMatrix = corMatrix, include.all=TRUE)
MS2_with_mz[1]
}
}
