% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFTfunctions.r
\name{group_FT.FUN}
\alias{group_FT.FUN}
\title{find overlap peaks in a mzmin-sorted data frame}
\usage{
group_FT.FUN(i, df, absMz = 0.01, absRt = 15)
}
\arguments{
\item{i}{numeric, row index, compare ith peak with (i+)th peaks}

\item{df}{data.frame or matrix, each row is a peak area, colnames are c("mzmin","mzmax","rtmin","rtmax"), element must be numeric, df is sorted by "mzmin"}

\item{absMz, absRt}{used for \code{\link{judgePeakOverlap}}}
}
\value{
a logical vector
}
\description{
group_FT.FUN is to compare ith peaks with peaks after it in a peakTable to find overlap peaks
}
