% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFTfunctions.r
\name{judgePeakOverlap}
\alias{judgePeakOverlap}
\title{Whether peak area overlap}
\usage{
judgePeakOverlap(ref, alt, Mz = TRUE, Rt = TRUE, absMz = 0,
  absRt = 15)
}
\arguments{
\item{ref, alt}{a numeric vector, orderd in "mzmin","mzmax","rtmin","rtmax"}

\item{Mz}{logical, whether judge m/z overlap}

\item{Rt}{logical, whether judge rt overlap}

\item{absMz}{tolerance of molecular weight difference in Da unit between ref m/z and alt m/z}

\item{absRt}{tolerance of retention time difference between ref rt and alt rt}
}
\value{
logical, indicating whether ref and alt peaks are overlap
}
\description{
Function judegPeakOverlap is used to judge whether two peaks ([mzmin, mzmax, rtmin, rtmax]) are overlap
}
\examples{
ref = c(70.001, 70.003, 26, 40)
alt = c(70.002, 70.004, 50,60)
judgePeakOverlap(ref=ref, alt=alt, Mz=TRUE, Rt=TRUE, absMz=0, absRt=15) 

}
