% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFTfunctions.r
\name{match.one2one}
\alias{match.one2one}
\title{feature unique match}
\usage{
match.one2one(M1_to_M2, M2_to_M1)
}
\arguments{
\item{M1_to_M2}{results from get_compareFT, M1 feature table mapping to M2 feature table}

\item{M2_to_M1}{reverse of M1_to_M2}
}
\value{
data frame
}
\description{
feature unique match
}
