% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2SimilarityFunctions.r
\name{MatchSpec}
\alias{MatchSpec}
\title{group m/z based on m/z tolerance}
\usage{
MatchSpec(spec, ppm.ms2match = 30, mz.ppm.thr = 400)
}
\arguments{
\item{spec}{numeric matrix, columns are "mz" and "intensity", represents an MS2 spectrum}

\item{ppm.ms2match}{numeric, m/z tolerance of MS2 in ppm mode}

\item{mz.ppm.thr}{used in GetDiffMZppm function}
}
\value{
MS2 spectrum
}
\description{
As MassSpectrometry has fluctuate in m/z detection, one ion can be detected as two similar m/z, thus we need to combine these two m/z.
After Combining, m/z is the mean, and intensity is the sum.
}
