% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MS2_spec}
\alias{MS2_spec}
\title{MS2 spectrum resulting from getMS2FromFiles function}
\format{A list containing two elements
\describe{
  \item{MS2}{MS2 spectrum}
  \item{presursorInfo}{precursor information corresponding MS2 spectrum}
}}
\source{
data("MS2_spec", package="RFQI")
}
\usage{
MS2_spec
}
\description{
MS2 spectrum resulting from getMS2FromFiles function
}
\keyword{datasets}
