% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFTfunctions.r
\docType{class}
\name{ParamSet-class}
\alias{ParamSet-class}
\alias{ParamSet}
\title{parameter set for peak detection and peaks group}
\description{
Objects of the type \code{ParamSet} allow xcms to find peaks
}
\section{Slots}{

\describe{
\item{\code{peakwith}}{numeric(2): vector defining peakwith in Centwave method}

\item{\code{ppm}}{numeric(1): numeric defining ppm in Centwave method, usually MassSpectrometry ppm}

\item{\code{noise}}{numeric(1): default is 0}

\item{\code{absMz}}{numeric(1)}

\item{\code{absRt}}{numeric(1)}

\item{\code{binSize}}{numeric(1): used for alignment}
}}

