% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateFTFunctions.r
\name{plotFeatureMap}
\alias{plotFeatureMap}
\title{Function plotFeatureMap is to plot Feature, peaks, and MS2 point in a figure}
\usage{
plotFeatureMap(idx, MS2_inner_cor, features, MS2DB, heatmap = FALSE)
}
\arguments{
\item{idx}{index}

\item{MS2_inner_cor}{MS2 similarity matrix for MS2 clustering}

\item{features}{containing peaks and features}

\item{MS2DB}{db.MS2 object}

\item{heatmap}{logical, if true, plot MS2 similarity heatmap, else plot featureMap}
}
\value{
figure
}
\description{
Function plotFeatureMap is to plot Feature, peaks, and MS2 point in a figure
}
