% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plotMS2Cluster}
\alias{plotMS2Cluster}
\title{plot MS2 Cluster}
\usage{
plotMS2Cluster(df, cutoff = 0.4, main = "", label = TRUE)
}
\arguments{
\item{df}{correlation matrix}

\item{cutoff}{similarity score threshold}

\item{main}{main title of plot}

\item{lable}{if TRUE, add similarity score on edges}
}
\description{
plot MS2 similarity network using igraph
}
