% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spectrumDB}
\alias{spectrumDB}
\title{standard metabolite MS2 spectrum}
\format{
A list containing two elements
\describe{
  \item{spectrum}{standard MS2 spectrum, each spectrum is a list, whose name is the metabolite ID, and value is spectrum}
  \item{Info}{metabolite information, including Labid, HMDBID, mz, formula, and HMDB name}
}
}
\source{
data("spectrumDB", package="RFQI")
}
\usage{
spectrumDB
}
\description{
standard metabolite MS2 spectrum
}
\keyword{datasets}
