% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateFTFunctions.r
\name{summary_L}
\alias{summary_L}
\title{summary_L is to compute rt and ppm shift}
\usage{
summary_L(df)
}
\arguments{
\item{df}{data frame whose colnames containing c("mzmin","mzmax","rtmin","rtmax")}
}
\value{
numeric matrix containing m/z difference, rt difference, and m/z difference in ppm unit
}
\description{
summary_L is to compute rt and ppm shift
}
