% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFTfunctions.r
\name{getPeakRange_singFile}
\alias{getPeakRange_singFile}
\title{group overlap peaks in a peak table
getPeakRange_singleFiles is to group peaks from a peaks table}
\usage{
getPeakRange_singFile(peaks, cores = 2, param = new("ParamSet"))
}
\arguments{
\item{peaks}{matrix}

\item{cores}{numeric, for parallel computing}

\item{param}{ParamSet object}
}
\value{
list
}
\description{
group overlap peaks in a peak table
getPeakRange_singleFiles is to group peaks from a peaks table
}
