% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclapply_hack.r
\name{mclapply}
\alias{mclapply}
\title{mclapply}
\usage{
mclapply(
  X,
  FUN,
  ...,
  mc.preschedule = TRUE,
  mc.set.seed = TRUE,
  mc.silent = FALSE,
  mc.cores = getOption("mc.cores", 2L),
  mc.cleanup = TRUE,
  mc.allow.recursive = TRUE,
  affinity.list = NULL
)
}
\description{
If the OS is Windows, set mclapply to the
the hackish version. Otherwise, leave the
definition alone.
}
