% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateFTFunctions.r
\name{plotArea}
\alias{plotArea}
\title{Function plotArea is used to plot feature area of one FT or two}
\usage{
plotArea(refPeaks, group = NULL, main = "", mzr = NULL, rtr = NULL)
}
\arguments{
\item{refPeaks}{dataframe or matrix containing columns of c("mzmin","mzmax","rtmin","rtmax")}

\item{group}{factor, length as row of refPeaks, indicating every features' group}

\item{main}{title of plot}

\item{mzr}{m/z range}

\item{rtr}{rt range}
}
\value{
plot
}
\description{
Function plotArea is used to plot feature area of one FT or two
}
